application/pyproject.toml
[tool.poetry]
name = "application"
version = "0.1.0"
description = ""
authors = ["Смелов Илья <https://github.com/PivnoyFei>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
django-cleanup = "^8.0.0"
django-debug-toolbar = "^4.1.0"
django-rest-framework = "^0.1.0"
djoser = "^2.2.0"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
pillow = "^9.5.0"
psycopg2-binary = "^2.9.6"
sorl-thumbnail = "^12.9.0"
pytils = "^0.4.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"

[tool.isort]
profile = "black"
line_length = 88
wrap_length = 88
skip = [".gitignore", ".dockerignore", "migrations"]
extend_skip = ["_version.py", "versioneer.py", "__init__.py"]
extend_skip_glob = ["docs/*", "*venv*/*"]

[tool.black]
skip-string-normalization = true
line-length = 88
extend-exclude = '(.*\/_version.py|.*venv.*\/|.*migrations.*\/.*)'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
